order completion

+++ app/controllers/orders.rb
@@ -6,7 +6,7 @@ class OrdersController < UsefulMusic::App
 
   def create
     shopping_basket_record = ShoppingBasket::Record[request.POST['order']['shopping_basket_id']]
-    Order::Record.last.destroy
+    Order::Record.last.destroy if Order::Record.last
     order = Order.create do |order|
       order.record.shopping_basket_record = shopping_basket_record
       order.record.state = 'initialized'
@@ -37,7 +37,7 @@ class OrdersController < UsefulMusic::App
         :category => :Digital
       }],
       :custom_fields => {
-        CARTBORDERCOLOR: "C000C0",
+        CARTBORDERCOLOR: "01344C",
         LOGOIMG: "https://example.com/logo.png"
       }
     )
@@ -58,14 +58,42 @@ class OrdersController < UsefulMusic::App
       :signature  => ENV.fetch('PAYPAL_SIGNATURE')
     )
     token = request.GET['token']
-    ap express_response = express_request.details(token)
-    ap express_response.payer.identifier
-    ap express_response.payer.first_name
-    ap express_response.payer.last_name
-    ap express_response.payer.email
-    ap express_response.payer.company
-    ap express_response.amount
-    ap express_response.ship_to
-    ap express_response.payment_responses
+    # ap express_response = express_request.details(token)
+    # ap express_response.payer.identifier
+    # ap express_response.payer.first_name
+    # ap express_response.payer.last_name
+    # ap express_response.payer.email
+    # ap express_response.payer.company
+    # ap express_response.amount
+    # ap express_response.ship_to
+    # ap express_response.payment_responses
+    payer_id = request.GET['PayerID']
+    payment_request = Paypal::Payment::Request.new(
+      :currency_code => :GBP,   # if nil, PayPal use USD as default
+      # :description   => 'Lovely',    # item description
+      # :quantity      => 1,      # item quantity
+      # :amount        => order.record.shopping_basket_record.purchase_records.map{|r| Purchase.new r}.reduce(0){|t, i| i.price},   # item value
+      :amount => 2.64,
+      :items => [{
+        :name => 'blah',
+        :description => 'some',
+        :amount => 1.32,
+        :quantity => 2,
+        :category => :Digital
+      }],
+      :custom_fields => {
+        CARTBORDERCOLOR: "C000C0",
+        LOGOIMG: "https://example.com/logo.png"
+      }
+    )
+    ap payer_id
+    ap token
+    checkout_response = express_request.checkout!(
+      token,
+      payer_id,
+      payment_request
+    )
+    ap checkout_response
+    
   end
 end
